#1 checkpoint

+++ src/AverageUser.java
@@ -8,7 +8,7 @@ public class AverageUser extends Account{
 	}
 	
 	public AverageUser(String id, String email, String name, String type, boolean isFlagged, long maxCheckout,
-			double balance, String passwordString, int age) {
+			double balance, String passwordString, long age) {
 		super(id, email, name, type, isFlagged, maxCheckout,
 				balance, passwordString, age);
 	}
